Adjustments to the API craft fetching script for when the comp doesn't exist or you want to keep the temporary files.

+++ _Other Stuff/get_craft.py
@@ -1,6 +1,7 @@
 # Standard library imports
 import argparse
 import json
+import os
 import re
 import subprocess
 import tempfile
@@ -10,6 +11,7 @@ from shutil import which
 parser = argparse.ArgumentParser(description="Grab all the craft from the API for a competition and rename them.", formatter_class=argparse.ArgumentDefaultsHelpFormatter)
 parser.add_argument('compID', type=int, help="Competition ID")
 parser.add_argument('--curl-command', type=str, default='curl', help="Curl command (in case it needs specifying on Windows).")
+parser.add_argument('--keep', action='store_true', help="Keep the player and manifest files.")
 args = parser.parse_args()
 
 if which(args.curl_command) is None:
@@ -18,8 +20,9 @@ if which(args.curl_command) is None:
 
 comp_url = f"https://conquertheair.com/competitions/{args.compID}/vessels/manifest.json"
 players_url = f"https://conquertheair.com/players"
-manifest_file = Path(tempfile.mkstemp()[1])
-players_file = Path(tempfile.mkstemp()[1])
+cwd = Path('.').absolute()
+manifest_file = Path(tempfile.mkstemp(dir=cwd if args.keep else None)[1])
+players_file = Path(tempfile.mkstemp(dir=cwd if args.keep else None)[1])
 
 try:
     print("Fetching players and competition manifest...", end='', flush=True)
@@ -38,21 +41,30 @@ try:
 
     with open(manifest_file, 'r') as f:
         manifest = json.load(f)
-        print(f"Fetching {len(manifest)} craft files...", end='', flush=True)
-        for craft in manifest:
-            craft_name = f"{playerIDs[craft['player_id']]}_{craft['name']}"
-            subprocess.run([args.curl_command, "-s", f"{craft['craft_url']}", "-o", f"{craft_name}.craft"])
-            with open(f"{craft_name}.craft", 'r+') as f:
-                lines = f.read().splitlines()
-                lines[0] = f"ship = {craft_name}"
-                f.seek(0, 0)
-                f.truncate()
-                f.write('\n'.join(lines))
-            print('.', end='', flush=True)
-        print("done.")
+        if "status" in manifest and manifest["status"] == 404:
+            print(f"No competition with ID {args.compID}")
+        else:
+            print(f"Fetching {len(manifest)} craft files...", end='', flush=True)
+            for craft in manifest:
+                craft_name = f"{playerIDs[craft['player_id']]}_{craft['name']}".replace(os.path.sep, '_')
+                subprocess.run([args.curl_command, "-s", f"{craft['craft_url']}", "-o", f"{craft_name}.craft"])
+                with open(f"{craft_name}.craft", 'r+') as f:
+                    lines = f.read().splitlines()
+                    lines[0] = f"ship = {craft_name}"
+                    f.seek(0, 0)
+                    f.truncate()
+                    f.write('\n'.join(lines))
+                print('.', end='', flush=True)
+            print("done.")
 
 finally:
-    if manifest_file.exists():
-        manifest_file.unlink()
-    if players_file.exists():
-        players_file.unlink()
+    if not args.keep:
+        if manifest_file.exists():
+            manifest_file.unlink()
+        if players_file.exists():
+            players_file.unlink()
+    else:
+        if manifest_file.exists():
+            manifest_file.rename('manifest.json')
+        if players_file.exists():
+            players_file.rename('players.html')
